Safari-609.1.11.1/LayoutTests/media/modern-media-controls/button/button-on-expected.txt

Testing an Button has a white background-color when set to `on`.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(button.image.element).backgroundColor is "rgb(255, 255, 255)"

PASS successfullyParsed is true

TEST COMPLETE
